requirements_final.txt
absl-py==2.0.0
addict==2.4.0
albumentations==0.4.6
annotated-types==0.6.0
anyio==3.7.1
astroid==2.5
astunparse==1.6.3
attrdict==2.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==4.2.4
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.9
contourpy==1.1.1
cycler==0.12.1
deepface==0.0.80
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==2.0.7
fonttools==4.46.0
fsspec==2023.12.1
future==0.18.3
gast==0.4.0
gdown==4.7.1
glog==0.3.1
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
imutils==0.5.4
iniconfig==2.0.0
isort==5.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jstyleson==0.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
libclang==16.0.6
lit==17.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.3.2
mccabe==0.6.1
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omz-tools @ file:///home/prixa-ml/Desktop/projects/liveness-detection/Face-Recognition/open_model_zoo/tools/model_tools
onnx==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openvino==2023.2.0
openvino-dev==2023.2.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
packaging==23.2
pandas==1.1.2
pep517==0.13.1
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pluggy==0.13.1
protobuf==3.20.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pylint==2.6.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==6.1.0
pytest-pylint==0.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gflags==3.1.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requirementslib==3.0.0
retina-face==0.0.13
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.7.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==2.0.0+cu117
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.12.1
yacs==0.1.8
yarg==0.1.9
zipp==3.17.0
